Update math and Solenoid

+++ include/ck_utilities/Joystick.hpp
@@ -12,7 +12,7 @@ public:
     Joystick(uint joystickID);
     static void update(const rio_control_node::Joystick_Status& joystick_status_msg);
     double getRawAxis(uint axisID);
-    double getFilteredAxis(uint axisID, float deadband);
+    double getFilteredAxis(uint axisID, double deadband);
     bool getAxisActuated(uint axisID, float threshold);
     bool getButton(uint buttonID);
     bool getRisingEdgeButton(uint buttonID);
